Echo current state being executed in deplo-eos script
Fixing a typo in deploy-eos script.
In the for loop for running salt states for ha.haproxy, misc.openldap
and the s3server components the echo needs to print the correct state
being run, for all above states it only prints 'running s3server state
on both the nodes' which gives an impression that the other two states
were not run.